Renamed package and return pointer now
